Resources Page Map Feature (#41)
* Working map that updates to show the user's location

* Update ResourcesMap.tsx

* Update ResourcesMap.tsx

* doesn't work great but the api collects cafes nearby

* Update ResourcesMap.tsx

* finally got markers to work in some capacity( kinda bugged)

* Marker and modal pop up

So the markers can be customized but I commented it out for now.  The modal pops up with data from the selected marker. My next hurdle to cross is to have the selectable chips actually work and to have custom icons/colors.

* chips work

The chips work, but now I need to have the chips update the map when selected as well as remove or add when they are deselected

* custom icons and removed scroll view and expanded map

* small update

* Update ResourcesMap.tsx

got the markers to show up as the buttons are pressed

* ui updates and now the map updates when the button is pressed

* If buttons are pressed and then all deselected the map markers will get wiped

* FULLY FUNCTIONAL!!! (not really dynamic but don;t worry about it ;))

* Update ResourcesMap.tsx

* added a get directions feature that redirects to google maps